demo-project/requirements.txt
google-genai
chromadb
GitPython
python-dotenv
sentence-transformers